requirements.txt
# ============================================
# OpenTutor Python Dependencies
# ============================================
# Python >= 3.10 required
# Install: pip install -r requirements.txt

# ============================================
# Core dependencies
# ============================================
python-dotenv>=1.0.0
PyYAML>=6.0
tiktoken>=0.5.0

# ============================================
# HTTP and API clients
# ============================================
requests>=2.31.0
openai>=1.0.0
perplexityai>=0.1.0
dashscope>=1.14.0
aiohttp>=3.9.0
httpx>=0.25.0

# ============================================
# Async support
# ============================================
nest_asyncio>=1.5.8

# ============================================
# Web framework and server
# ============================================
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ============================================
# RAG and knowledge base
# ============================================
lightrag-hku>=1.0.0
raganything>=0.1.0

# ============================================
# Academic and research tools
# ============================================
arxiv>=2.0.0

# ============================================
# Scientific computing (optional, for code execution)
# ============================================
# Uncomment if users need these for code execution feature
# numpy>=1.24.0
# matplotlib>=3.7.0
# scipy>=1.11.0
# pandas>=2.0.0

# ============================================
# Development tools
# ============================================
pre-commit>=3.0.0
